requirements.txt
Django==2.2.16
djangorestframework==3.12.4
pytz==2022.1
sqlparse==0.4.2
djoser
django-filter==22.1
Pillow==9.2.0
drf-extra-fields==3.4.0
